Changed the profit to be the key, won't produce results if no max profit item available.

+++ src/main/java/onesilver500buy/OneSilver500BuyJob.java
@@ -3,7 +3,7 @@ package onesilver500buy;
 import gw2.GW2Writable;
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.fs.Path;
-import org.apache.hadoop.io.IntWritable;
+import org.apache.hadoop.io.DoubleWritable;
 import org.apache.hadoop.mapreduce.Job;
 import org.apache.hadoop.mapreduce.lib.input.FileInputFormat;
 import org.apache.hadoop.mapreduce.lib.output.FileOutputFormat;
@@ -16,7 +16,7 @@ public class OneSilver500BuyJob {
         job.setMapperClass(OneSilver500BuyMapper.class);
         job.setCombinerClass(OneSilver500BuyReducer.class);
         job.setReducerClass(OneSilver500BuyReducer.class);
-        job.setOutputKeyClass(IntWritable.class);
+        job.setOutputKeyClass(DoubleWritable.class);
         job.setOutputValueClass(GW2Writable.class);
 
         FileInputFormat.addInputPath(job, new Path(args[0]));
